@@ -3,6 +3,7 @@
   "version": "0.0.1",
   "description": "Starter responsive dashboard layout built with tailwindcss and vue 3",
   "license": "MIT",
+  "type": "module",
   "keywords": [
     "dashboard",
     "tailwindcss",
@@ -23,18 +24,18 @@
     "deploy": "gh-pages -d dist"
   },
   "dependencies": {
-    "@headlessui/vue": "^1.4.1",
-    "@heroicons/vue": "^1.0.4",
-    "chart.js": "^3.5.1",
-    "vue": "^3.0.5",
-    "vue-router": "4"
+    "@headlessui/vue": "^1.7.23",
+    "@heroicons/vue": "1",
+    "chart.js": "3.5.*",
+    "vue": "^3.5.13",
+    "vue-router": "^4.5.0"
   },
   "devDependencies": {
-    "@vitejs/plugin-vue-jsx": "^1.1.7",
-    "autoprefixer": "^10.3.3",
-    "gh-pages": "^3.1.0",
-    "postcss": "^8.3.6",
-    "tailwindcss": "^2.2.9",
-    "vite": "2.4.2"
+    "@vitejs/plugin-vue-jsx": "^4.1.1",
+    "autoprefixer": "^10.4.20",
+    "gh-pages": "^6.3.0",
+    "postcss": "^8.5.1",
+    "tailwindcss": "3.*",
+    "vite": "^6.0.11"
   }
 }
